requirements.txt
Flask
Flask-SQLAlchemy
Flask-Login
psycopg2-binary
python-dotenv
Werkzeug
SQLAlchemy
gunicorn
supabase